extract common strategy functionality into mixins
